refactor: client와 server 컴포넌트 분리
